much more reliable scan results
Improved the interrupt service routine by stripping it down to the real basics.
No false positives by decodeing errors anymore.

+++ Funkthermometer/WTDecoder.h
@@ -24,7 +24,7 @@ public:
     byte channel;       // channel switch of transmitter 1,2,3
     int16_t temprature; // terperature in 1/10 °C
     int16_t humidity;   // humidity in %
-    unsigned long timestamp;
+    uint32_t timestamp; // time this record was created
   };
 
   WTDecoder() {};
